requirements.txt
appnope==0.1.0
decorator==4.0.6
future==0.15.2
fuzzywuzzy==0.8.0
gnureadline==6.3.3
ipython==4.0.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pymongo==3.2
python-Levenshtein==0.12.0
python-telegram-bot==3.1.1
simplegeneric==0.8.1
traitlets==4.0.0
Twisted==15.5.0
zope.interface==4.1.3
